Please merge because my repository was renamed from SolrIndexChangeAlerttxtx to SolrIndexChangeAlert (#1)
* Added missing dependency for log4j in build.gradle. Also turned off some logging in
IndexChangeListenerImpl.java

* Removed HEADER_MSG from classes as it is not needed by logfj.

* Added a TODO list to the README.txt file.

TODO:

   Simplify the Front End UI by getting rid of the radio buttons for
turning on/off
the index file monitoring. This action will then be incorporated into
the
WebSocket connection: connect switch.

   Add a properties file to replace com.prototype.utils.AppConstants so
that recompliling
will not be necessary.

* Debuggin was done. Fixed some JavaScript logic with radio buttons: start
stop. Converted the StompMessageClient to a Singleton as it should be.
Other changes were formating fixes.

* Modified README.txt

* Fixed the Singleton class StompMessageClient - was not implemeted
correctly.

* Design change of the Front End UI. Removed the start stop radio
buttons for the Index Monitor thread. This is now started and
stopped concurrently with starting and stopping the WebSocket connection.